+++ src/libAtomVM/otp_socket.c
@@ -67,6 +67,23 @@
 
 #define TAG "otp_socket"
 
+#if OTP_SOCKET_BSD
+static inline bool errno_is_peer_close(int err)
+{
+    if (err == ECONNRESET) {
+        return true;
+    }
+#ifdef __wasip2__
+    // On wasm32-wasip2, wasi-libc maps wasi:sockets@0.2.x stream-end to
+    // EIO on read and EPIPE on write
+    if (err == EIO || err == EPIPE) {
+        return true;
+    }
+#endif
+    return false;
+}
+#endif
+
 // Check some LWIP options
 #if OTP_SOCKET_LWIP
 #if !TCP_LISTEN_BACKLOG
@@ -214,7 +231,7 @@ static const char *const reuseaddr_atom = ATOM_STR("\x9", "reuseaddr");
 static const char *const type_atom = ATOM_STR("\x4", "type");
 static const char *const add_membership_atom = ATOM_STR("\xE", "add_membership");
 
-#define CLOSED_FD 0
+#define CLOSED_FD (-1)
 
 #define ADDR_ATOM globalcontext_make_atom(global, addr_atom)
 #define CLOSE_INTERNAL_ATOM globalcontext_make_atom(global, close_internal_atom)
@@ -648,9 +665,8 @@ static term nif_socket_open(Context *ctx, int argc, term argv[])
 #endif
 #if OTP_SOCKET_BSD
     rsrc_obj->fd = socket(domain, type, protocol);
-    if (UNLIKELY(rsrc_obj->fd == -1 || rsrc_obj->fd == CLOSED_FD)) {
+    if (UNLIKELY(rsrc_obj->fd == CLOSED_FD)) {
         AVM_LOGE(TAG, "Failed to initialize socket.");
-        rsrc_obj->fd = CLOSED_FD;
         enif_release_resource(rsrc_obj);
         return make_errno_tuple(ctx);
     } else {
@@ -811,7 +827,7 @@ static term nif_socket_close(Context *ctx, int argc, term argv[])
     SMP_RWLOCK_WRLOCK(rsrc_obj->socket_lock);
 
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd) {
+    if (rsrc_obj->fd != CLOSED_FD) {
         // In POSIX with BSD sockets, if a file descriptor being monitored by
         // select() is closed in another thread, the result is unspecified.
         // select may continue.
@@ -1396,7 +1412,7 @@ static term nif_socket_getopt(Context *ctx, int argc, term argv[])
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
 
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state == SocketStateClosed) {
 #endif
@@ -1485,7 +1501,7 @@ static term nif_socket_setopt(Context *ctx, int argc, term argv[])
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
 
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state == SocketStateClosed) {
 #endif
@@ -1508,6 +1524,10 @@ static term nif_socket_setopt(Context *ctx, int argc, term argv[])
                 int res = setsockopt(rsrc_obj->fd, SOL_SOCKET, SO_REUSEADDR, &option_value, sizeof(int));
                 SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
                 if (UNLIKELY(res != 0)) {
+                    // SO_REUSEADDR may be unavailable (ENOPROTOOPT)
+                    if (errno == ENOPROTOOPT) {
+                        return OK_ATOM;
+                    }
                     return make_errno_tuple(ctx);
                 } else {
                     return OK_ATOM;
@@ -1537,6 +1557,7 @@ static term nif_socket_setopt(Context *ctx, int argc, term argv[])
                 VALIDATE_VALUE(linger, term_is_integer);
 
 #if OTP_SOCKET_BSD
+#ifdef SO_LINGER
                 struct linger sl;
                 sl.l_onoff = (onoff == TRUE_ATOM);
                 sl.l_linger = term_to_int(linger);
@@ -1547,6 +1568,12 @@ static term nif_socket_setopt(Context *ctx, int argc, term argv[])
                 } else {
                     return OK_ATOM;
                 }
+#else
+                // Silently accept SO_LINGER if underlying doesn't know about it
+                UNUSED(onoff);
+                SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
+                return OK_ATOM;
+#endif
 #elif OTP_SOCKET_LWIP
                 rsrc_obj->linger_on = (onoff == TRUE_ATOM);
                 rsrc_obj->linger_sec = term_to_int(linger);
@@ -1693,12 +1720,12 @@ static term nif_socket_sockname(Context *ctx, int argc, term argv[])
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
 
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state == SocketStateClosed) {
 #endif
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 
 #if OTP_SOCKET_BSD
@@ -1765,14 +1792,14 @@ static term nif_socket_peername(Context *ctx, int argc, term argv[])
 
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state == SocketStateClosed) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
     if (rsrc_obj->socket_state & SocketStateUDP) {
         // TODO: handle "connected" UDP sockets
@@ -1843,14 +1870,14 @@ static term nif_socket_bind(Context *ctx, int argc, term argv[])
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
 #if OTP_SOCKET_BSD
     TRACE("rsrc_obj->fd=%i\n", (int) rsrc_obj->fd);
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state == SocketStateClosed) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 #endif
 
@@ -1940,8 +1967,6 @@ static term nif_socket_listen(Context *ctx, int argc, term argv[])
     TRACE("nif_socket_listen\n");
     UNUSED(argc);
 
-    GlobalContext *global = ctx->global;
-
     VALIDATE_VALUE(argv[0], term_is_otp_socket);
     VALIDATE_VALUE(argv[1], term_is_integer);
 
@@ -1952,18 +1977,18 @@ static term nif_socket_listen(Context *ctx, int argc, term argv[])
 
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state == SocketStateClosed) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
     if (rsrc_obj->socket_state & SocketStateUDP) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EPROTOTYPE, global), ctx);
+        return make_error_tuple(posix_errno_to_term(EPROTOTYPE, ctx->global), ctx);
     }
 #endif
 
@@ -2043,14 +2068,14 @@ static term nif_socket_accept(Context *ctx, int argc, term argv[])
 
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state & SocketStateClosed) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
     if (rsrc_obj->socket_state & SocketStateUDP) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
@@ -2073,7 +2098,7 @@ static term nif_socket_accept(Context *ctx, int argc, term argv[])
     }
     int fd = accept(rsrc_obj->fd, (struct sockaddr *) &clientaddr, &clientlen);
     SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-    if (UNLIKELY(fd == -1 || fd == CLOSED_FD)) {
+    if (UNLIKELY(fd == CLOSED_FD)) {
         int err = errno;
         if (err != EAGAIN) {
             AVM_LOGI(TAG, "Unable to accept on socket %i.  errno=%i", rsrc_obj->fd, (int) err);
@@ -2334,7 +2359,7 @@ static term nif_socket_recv_with_peek(Context *ctx, term resource_term, struct S
     if (res < 0) {
         if (errno == EAGAIN) {
             return make_error_tuple(TIMEOUT_ATOM, ctx);
-        } else if (errno == ECONNRESET) {
+        } else if (errno_is_peer_close(errno)) {
             TRACE("Peer closed connection.\n");
             return make_error_tuple(CLOSED_ATOM, ctx);
         }
@@ -2418,7 +2443,7 @@ static term nif_socket_recv_without_peek(Context *ctx, term resource_term, struc
         if (res < 0) {
             int err = errno;
             free(buffer);
-            if (err == ECONNRESET) {
+            if (errno_is_peer_close(err)) {
                 TRACE("Peer closed connection.\n");
                 return make_error_tuple(CLOSED_ATOM, ctx);
             } else if (err == EAGAIN) {
@@ -2568,14 +2593,14 @@ static term nif_socket_recv_internal(Context *ctx, term argv[], bool is_recvfrom
 
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, ctx->global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state & SocketStateClosed) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, ctx->global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
     if (rsrc_obj->socket_state & SocketStateListening) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
@@ -2616,9 +2641,12 @@ static term nif_socket_recvfrom(Context *ctx, int argc, term argv[])
 //
 // send/sendto
 //
-static ssize_t do_socket_send(struct SocketResource *rsrc_obj, const uint8_t *buf, size_t len, term dest)
+static ssize_t do_socket_send(struct SocketResource *rsrc_obj, const uint8_t *buf, size_t len, term dest, int *out_errno)
 {
     ssize_t sent_data = -1;
+    if (out_errno != NULL) {
+        *out_errno = 0;
+    }
 #if OTP_SOCKET_BSD
     if (!term_is_invalid_term(dest)) {
         struct RefcBinary *rsrc_refc = refc_binary_from_data(rsrc_obj);
@@ -2643,10 +2671,14 @@ static ssize_t do_socket_send(struct SocketResource *rsrc_obj, const uint8_t *bu
         sent_data = send(rsrc_obj->fd, buf, len, 0);
     }
     if (sent_data < 0) {
-        if (errno == EAGAIN || errno == EWOULDBLOCK) {
+        int err = errno;
+        if (out_errno != NULL) {
+            *out_errno = err;
+        }
+        if (err == EAGAIN || err == EWOULDBLOCK) {
             return SocketWouldBlock;
         }
-        if (errno == EBADF || errno == ECONNRESET) {
+        if (errno_is_peer_close(err)) {
             return SocketClosed;
         }
         return SocketOtherError;
@@ -2710,6 +2742,9 @@ static ssize_t do_socket_send(struct SocketResource *rsrc_obj, const uint8_t *bu
     }
     LWIP_END();
     if (err == ERR_CLSD) {
+        if (out_errno != NULL) {
+            *out_errno = EPIPE;
+        }
         return SocketClosed;
     }
     if (sent_data == 0 && err == ERR_MEM) {
@@ -2725,7 +2760,7 @@ static ssize_t do_socket_send(struct SocketResource *rsrc_obj, const uint8_t *bu
 ssize_t socket_send(struct SocketResource *rsrc_obj, const uint8_t *buf, size_t len, term dest)
 {
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
-    ssize_t result = do_socket_send(rsrc_obj, buf, len, dest);
+    ssize_t result = do_socket_send(rsrc_obj, buf, len, dest, NULL);
     SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
     return result;
 }
@@ -2747,14 +2782,14 @@ static term nif_socket_send_internal(Context *ctx, int argc, term argv[], bool i
 
     SMP_RWLOCK_RDLOCK(rsrc_obj->socket_lock);
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state == SocketStateClosed) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
     if (rsrc_obj->socket_state & SocketStateListening) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
@@ -2771,40 +2806,39 @@ static term nif_socket_send_internal(Context *ctx, int argc, term argv[], bool i
     const uint8_t *buf = (const uint8_t *) term_binary_data(data);
     size_t len = term_binary_size(data);
 
-    ssize_t sent_data = do_socket_send(rsrc_obj, buf, len, dest);
+    int send_errno = 0;
+    ssize_t sent_data = do_socket_send(rsrc_obj, buf, len, dest, &send_errno);
     SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
+
+    // An empty payload, such as an empty UDP datagram, sends 0 bytes, which is
+    // also SocketClosed: only a closed socket sets an errno.
+    if (len == 0 && sent_data == 0 && send_errno == 0) {
+        return OK_ATOM;
+    }
 
     // {ok, RestData} | {error, Reason}
 
-    // Transient send backpressure (lwIP ERR_MEM / BSD EAGAIN|EWOULDBLOCK) is
-    // reported as {error, eagain} so callers can retry rather than mistaking it
-    // for a closed connection.
-    if (sent_data == SocketWouldBlock) {
-        return make_error_tuple(posix_errno_to_term(EAGAIN, global), ctx);
-    }
-
-    size_t rest_len = len - sent_data;
-    if (rest_len == 0) {
-        return OK_ATOM;
-    } else if (sent_data > 0) {
-
+    if (sent_data > 0) {
+        size_t rest_len = len - (size_t) sent_data;
+        if (rest_len == 0) {
+            return OK_ATOM;
+        }
         size_t requested_size = term_sub_binary_heap_size(data, rest_len);
         if (UNLIKELY(memory_ensure_free_with_roots(ctx, TUPLE_SIZE(2) + requested_size, 1, &data, MEMORY_CAN_SHRINK) != MEMORY_GC_OK)) {
             AVM_LOGW(TAG, "Failed to allocate memory: %s:%i.", __FILE__, __LINE__);
             RAISE_ERROR(OUT_OF_MEMORY_ATOM);
         }
-
         term rest = term_maybe_create_sub_binary(data, sent_data, rest_len, &ctx->heap, global);
         return port_create_tuple2(ctx, OK_ATOM, rest);
-
-    } else if (sent_data == 0) {
-        // do_socket_send only returns 0 for a closed connection (SocketClosed:
-        // lwIP ERR_CLSD, BSD EBADF|ECONNRESET); an empty payload already
-        // returned ok above via the rest_len == 0 check. Report it as such.
-        return make_error_tuple(CLOSED_ATOM, ctx);
-    } else {
-        TRACE("Unable to send data: res=%zi.\n", sent_data);
-        return make_error_tuple(CLOSED_ATOM, ctx);
+    }
+    TRACE("Unable to send data: res=%zi.\n", sent_data);
+    switch (sent_data) {
+        case SocketWouldBlock:
+            return make_error_tuple(posix_errno_to_term(EAGAIN, global), ctx);
+        case SocketClosed:
+            return make_error_tuple(CLOSED_ATOM, ctx);
+        default:
+            return make_error_tuple(posix_errno_to_term(send_errno != 0 ? send_errno : EINVAL, global), ctx);
     }
 }
 
@@ -2969,14 +3003,14 @@ static term nif_socket_connect(Context *ctx, int argc, term argv[])
     }
 
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state == SocketStateClosed) {
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
     if (((rsrc_obj->socket_state & SocketStateTCPListening) == SocketStateTCPListening)
         || ((rsrc_obj->socket_state & SocketStateTCPConnected) == SocketStateTCPConnected)) {
@@ -3099,12 +3133,12 @@ static term nif_socket_shutdown(Context *ctx, int argc, term argv[])
     }
 
 #if OTP_SOCKET_BSD
-    if (rsrc_obj->fd == 0) {
+    if (rsrc_obj->fd == CLOSED_FD) {
 #elif OTP_SOCKET_LWIP
     if (rsrc_obj->socket_state == SocketStateClosed) {
 #endif
         SMP_RWLOCK_UNLOCK(rsrc_obj->socket_lock);
-        return make_error_tuple(posix_errno_to_term(EBADF, global), ctx);
+        return make_error_tuple(CLOSED_ATOM, ctx);
     }
 
     term result = OK_ATOM;